Rename workspaces feature to home and route home page to it
