+++ client/features/home/WorkspacePreview.tsx
