+++ client/features/workspace/WorkspaceScreen.tsx
@@ -17,7 +17,7 @@ import { McpMenu } from '@/client/features/connectors/McpMenu'
 import { WidgetErrorBoundary } from '@/client/features/applets/WidgetErrorBoundary'
 import { Widgets } from '@/client/features/widgets/Widgets'
 import { PanelHeader } from '@/client/components/shared/PanelHeader'
-import { workspaceProviderIcon } from '@/client/features/workspaces/workspace-presentation'
+import { workspaceProviderIcon } from '@/client/features/home/workspace-presentation'
 import { Button } from '@/client/components/ui/button'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/client/components/ui/tooltip'
 import { WorkspaceSettings } from '@/client/features/settings/WorkspaceSettings'
